Locks are applied with resolutionStrategy.dependencySubstitution.substitute instead of resolutionStrategy.eachDependency with useVersion to assist with core gradle alignment platform edges
Core Gradle alignment does not apply `useVersion`/`useTarget` to virtual platforms, which make up aligned groups. Instead, this more specific DSL of dependencySubstitution is used and applied to virtual platforms.

For the cases where there is a non-locked alignment rule that causes a resolution difference to that seen in the lock file, the resolved version will equal the locked version with Nebula alignment and resolution will fail due to `Multiple forces on different versions for virtual platform` with core Gradle alignment (causing the user to make a change or update their locks)

Without this change, this case above with core Gradle alignment would result in the locked version and the new aligned version entering conflict resolution and the locked version may no longer be the final resolved version.