DependencyResolutionVerifier does not verify that the resolved version is the locked version when there has been a selective dependency update in the same command
This is currently only run with using core Gradle alignment and Nebula locking.

The information about which locks should be considered for this validation can show a false-positive, such as when `dependencyReport` is run in the same task invocation as `updateDependency`. This is likely related to dependencies from a recommendation BOM or from aligned dependencies. We can disable this additional validation when running selective dependency lock updates to avoid seeing a false positive error.